Move list expression planning to expr module
